* Don't try to remove the todo directory (Bug #36989)
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-1@59720 1b283449-9f2f-0410-b571-8eb228e1a901

+++ base/univention-quota/quota.py
@@ -215,6 +215,8 @@ def clean():
 	try:
 		if os.path.exists(SHARE_CACHE_DIR):
 			for filename in os.listdir(SHARE_CACHE_DIR):
+				if filename == 'todo':
+					continue
 				os.remove(os.path.join(SHARE_CACHE_DIR, filename))
 		if os.path.exists(SHARE_CACHE_TODO_DIR):
 			for filename in os.listdir(SHARE_CACHE_TODO_DIR):
